backend/requirements.txt
Flask~=3.1.1
Flask-Migrate~=4.1.0
Flask-SQLAlchemy~=3.1.1
python-dotenv~=1.1.1
pyjwt~=2.10.1
requests~=2.32.4
psycopg2-binary~=2.9.10